made style for Browse calendars and Favourites, moved code from Favourites to Favourite
