+++ frontend/src/routes/CalendarsPage.tsx
@@ -5,9 +5,13 @@ const Favourite: React.FC = () => {
 
     const [search, setSearch] = useState('');
 
+    const handleSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
+      setSearch(e.target.value);
+    };
+
     return (
       <div className="home calendars">
-        <Calendars search={search} setSearch={setSearch}/>
+        <Calendars search={search} setSearch={setSearch} handleSearch={handleSearch}/>
       </div>
     );
   };
